changed my Invoice Component to read from Dynamics through ADAL
Created a PatientContextWrapper Component to wrap the HomePage and load the patient record from dynamics,
so as to pass the patient data down the component tree as props,
and make them availble to all our components as state.

Changed a couple lines in App to target new wrapper components.

Made a wrapper for my Corona Component so as to display both National and Local Data.

Not Done with State Component, but placeHolder component and structure is there.