Made the necessary changes to be able to display the users local State data in the Corona Component
This required splitting the API call function, so it is no longer a general purpose wrapper for the Corona API.

As such, there are now two sets of constants/reducers/actions/wrapper and display components; one for national and one for state data.

This results in some pretty ugly and cluttered looking code.  However, it seems to be the only way to get these 2 components to display
independently of each other.  Trying to do it any other way simply displays the national data twice.

Converted the Corona Panel Component from a simple functional component into a Redux Container Component,
so as to pass down the patient profile context established at the top level of the application as props.

This allows the CoronaState Component which is a 3rd tier container component to dynamically call based on the patient's state of residence.