Made the necessary changes to dynamically fetch user profiles from Dynamics, instead of hard coding will's record.
Note that, in order for the other components to work correctly, a patient record has to have an email address,
matching the email provided at registration on the portal.

Also, the address1_stateorprovince field must be an actual state abreviation code, and there must be at least one
invoice associated with that patient.