Fixed profile caching error... sort of
There was an error where logging out and logging back in with another user
would render the profile of the previous user.

The issue was two-fold.  First the correct state was not being rendered by the component.
This was resolved by wrapping it and passing the profile as props instead of doing useEffect.

However, the component even after recieving the correct data, verified in the console,
would not render the current profile, but would still render the old page.

One would think this is just a browser caching error.
However, this issue is present in both Firefox and Chrome, and both incognito and normal browsing modes.

I tried every combination of redux calls/hooks imaginable including React's forceUpdate(),
to get the profilepage to manually re-render, even this did not work.

The only way to get the page to display correctly was to manually click refresh,
either on the profilepage or the login page.

So, as a temporary workaround, I had the logout handler trigger a hard refresh ignoring cache.
On chrome, this works seamlessly.  On Firefox, the hard reload is a little more obvious.
This really seemed like the only place where  programatic hard reload could even make a little sense.

However, it is clear to me that the issue is with Redux.
So, I will continue to try to dig around in it and see what I can do about a more elegant solution.